# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: province
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{eventId}/join:
    post:
      operationId: joinEvent
      tags: [events]
      summary: Join an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}/leave:
    post:
      operationId: leaveEvent
      tags: [events]
      summary: Leave an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /map/markers:
    get:
      operationId: getMapMarkers
      tags: [map]
      summary: Get map markers
      parameters:
        - name: precision
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Markers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapMarker"
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        userPhotoUrl:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        eventDate:
          type: string
        eventTime:
          type: string
        endDate:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        participantCount:
          type: integer
        isParticipating:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - username
        - title
        - location
        - eventDate
        - eventTime
        - participantCount
        - isParticipating
        - createdAt
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          nullable: true
        location:
          type: string
          maxLength: 200
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        eventDate:
          type: string
        eventTime:
          type: string
        endDate:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
      required:
        - title
        - location
        - eventDate
        - eventTime
    UpdateEventBody:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          nullable: true
        location:
          type: string
          maxLength: 200
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        eventDate:
          type: string
        eventTime:
          type: string
        endDate:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - title
        - message
        - isRead
        - createdAt
...
    MapMarker:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        count:
          type: integer
        locationName:
          type: string
          nullable: true
        trees:
          type: array
          items:
            type: object
      required:
        - latitude
        - longitude
        - count
        - trees
...